requirements.txt
pyyaml>=6.0.1
loguru>=0.7.2
opencv-python>=4.8.0
uiautomator2>=2.16.0
python-adb>=0.1.6